static_demo/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
python-multipart==0.0.5
click==8.0.3
pillow==8.4.0
tensorflow>=2.7.0,<2.9.0
jinja2==3.0.3
opencv-python-headless==4.5.4.60
numpy
keras